Split cops
