+++ lib/rubocop/cop/betterment.rb
@@ -20,4 +20,5 @@ require 'rubocop/cop/betterment/server_error_assertion'
 require 'rubocop/cop/betterment/hardcoded_id'
 require 'rubocop/cop/betterment/vague_serialize'
 require 'rubocop/cop/betterment/fetch_boolean'
-require 'rubocop/cop/betterment/status_code'
+require 'rubocop/cop/betterment/render_status'
+require 'rubocop/cop/betterment/redirect_status'
