+++ spec/rubocop/cop/betterment/redirect_status_spec.rb
@@ -2,31 +2,7 @@
 
 require 'spec_helper'
 
-describe RuboCop::Cop::Betterment::StatusCode, :config do
-  it 'adds an offense when rendering without a status' do
-    expect_offense(<<~RUBY)
-      def create
-        render :new
-        ^^^^^^^^^^^ Did you forget to specify an HTTP status code? [...]
-      end
-
-      def update
-        render plain: 'OK'
-        ^^^^^^^^^^^^^^^^^^ Did you forget to specify an HTTP status code? [...]
-      end
-    RUBY
-
-    expect_correction(<<~RUBY)
-      def create
-        render :new, status: :unprocessable_entity
-      end
-
-      def update
-        render plain: 'OK', status: :unprocessable_entity
-      end
-    RUBY
-  end
-
+describe RuboCop::Cop::Betterment::RedirectStatus, :config do
   it 'adds an offense when redirecting without a status' do
     expect_offense(<<~RUBY)
       def create
@@ -54,28 +30,22 @@ describe RuboCop::Cop::Betterment::StatusCode, :config do
   it 'does not add offenses for valid usage' do
     expect_no_offenses(<<~RUBY)
       def index
-        render :new
         redirect_to '/'
       end
 
       def show
-        render :new
         redirect_to '/'
       end
 
       def new
-        render :new
         redirect_to '/'
       end
 
       def edit
-        render :new
         redirect_to '/'
       end
 
       def create
-        render plain: "OK", status: :ok
-        render :new, status: :unprocessable_entity
         redirect_to '/', status: :found
         redirect_to '/', status: :see_other
       end
